Write a new version of the chaffin search algorithm
The previous implement was _almost_ correct but I was getting some weird
results with later subproblems. Specifically, at waste=100 the algorithm
produced 542 as the solution when a string definitely exists that
contains 543 permutations. I tried to debug and understand what was
going on but in the end decided to rewrite the algorithm to be a bit
clearer, with comments, named variables, etc.

This approach no longer uses 16-bit integers to store max_perms_array
and it stores simple values rather than relative stack pointer depths.
This caused a lot of trouble during debugging and I don't think it
actually helps to pack these integers tightly into registers.

The motivation for packing the integers this way was so we could
retrieve all w=1..4 values in one go, but I realised we probably should
be re-checking if best_perms can actually be reached after each
permutation is visited. Otherwise, one of the search branches could have
meant this check would have started to fail and we're needlessly
searching regions of space that can't yield improvements.

That does mean going to memory (likely L1 cache) more times than
previously, but this _should_ be a minor slow down in comparison to the
branches we'd eliminate from the search tree (I think).

Finally, we can make clever use of shifting when loading/storing values
to save a few instructions and some nice things like 'cbnz' also help.