Allow the search to continue from a previously known point
If we’ve run the search before, we can seed the search with the array of
known maximum permutations for W wasted characters to resume from that
point.

This also lets us guess what the next solution(s) might be and start
working on the next subproblem while waiting for the current iteration
to finish.